Set Testmode default to True
The blogpost [Set Intune Primary User with Azure Automation](https://www.tbone.se/2025/03/31/set-intune-primary-user-with-azure-automation/) states that $ExecutionMode is set to "Test" or "Prod". 
I presumed that you replaced this string param $ExecutionMode with the bool param $Testmode. I updated the HelpMessage and added a hint to the old param name and escaped it.
That parameter defaulted to $false. I updated the default value to $true as stated in the HelpMessage.

Please review. Thanks, Markus